fix loading, add appbar, reorganize frontend
